CSV? JSON?
It works well with json file. However according to Vienno’s video, it
won’t affect the web display if I change json data file to csv data
file. But in my case, the page displays nothing after replacing the
data source with mydata.csv…

Need to check it further.

And another question: when loading external data sets, is it exactly
the same when using the csv and json files??